VTCORE-VS-107 Rename nlppp project to nlp
