+++ nlp/nlp.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "resource.h"
+
+const _TCHAR *STRNULL = _T("");
