--- nlppp/nlppp.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include "resource.h"
-
-const _TCHAR *STRNULL = _T("");
